Rename 'utils' to 'objects'
Could had been a sub-module, but think this should be fine for now.
