+++ packages/support/src/objects/has.ts
