+++ packages/support/src/objects/hasAll.ts
