+++ packages/support/src/objects/hasAny.ts
